PR comment: rename modal file
