+++ src/js/common/index.js
@@ -3,7 +3,7 @@ import './polyfills';
 import './sentry.js';
 import '../legacy/menu';  // Used in the footer.
 import './usa-banner-toggle';
-import './utils/accessible-modal';
+import './utils/accessible-VCL-modal';
 import addMenuListeners from './utils/accessible-menus';
 
 addMenuListeners(document.querySelector('#vetnav-menu'), true);
