+++ src/js/common/utils/accessible-VCL-modal.js
